Add inline login/register flows, fix register bugs
