+++ src/components/molecules/form/checkbox.module.scss
@@ -6,7 +6,7 @@
 	margin-bottom: 16px;
 }
 
-.checkbox {
+.checkbox.checkbox {
 	-webkit-appearance: none;
 	appearance: none;
 	position: relative;
