+++ src/components/templates/andGlobalModals.tsx
@@ -2,16 +2,19 @@ import clsx from 'clsx';
 import dynamic from 'next/dynamic';
 import React, { ReactNode, useState } from 'react';
 import { FormattedMessage } from 'react-intl';
+import { useQueryClient } from 'react-query';
 
 import Button from '@components/molecules/button';
-import { makeLoginRoute, makeRegisterRoute } from '@lib/routes';
+import LoginForm from '@components/molecules/loginForm';
+import SocialLogin from '@components/molecules/socialLogin';
+import RegisterForm from '@components/organisms/registerForm';
+import { refetchUserQueries } from '@lib/api';
 import { UnreachableCaseError } from '@lib/typeHelpers';
-import useLanguageRoute from '@lib/useLanguageRoute';
 
 import styles from './andGlobalModals.module.scss';
 
 export type GlobalModalsContextType = {
-	challengeAuth: () => void;
+	challengeAuth: (onConfirm: () => unknown) => void;
 	confirmRemoveFavorite: (onConfirm: () => unknown) => void;
 };
 
@@ -31,7 +34,7 @@ interface AndGlobalModalsProps {
 export default function AndGlobalModals({
 	children,
 }: AndGlobalModalsProps): JSX.Element {
-	const languageRoute = useLanguageRoute();
+	const queryClient = useQueryClient();
 	const [isOpen, setIsOpen] = useState(false);
 	const [whichModal, setWhichModal] = useState<
 		'auth' | 'confirmRemoveFavorite'
@@ -39,15 +42,20 @@ export default function AndGlobalModals({
 	const [onConfirm, setOnConfirm] = useState<undefined | (() => unknown)>(
 		undefined
 	);
+	const [authState, setAuthState] = useState<'intro' | 'register' | 'login'>(
+		'intro'
+	);
 
 	const handleClose = () => {
 		setIsOpen(false);
 	};
 
 	const globalModalsContext: GlobalModalsContextType = {
-		challengeAuth: () => {
+		challengeAuth: (callback) => {
 			setWhichModal('auth');
 			setIsOpen(true);
+			setOnConfirm(() => callback);
+			setAuthState('intro');
 		},
 		confirmRemoveFavorite: (callback) => {
 			setWhichModal('confirmRemoveFavorite');
@@ -60,55 +68,102 @@ export default function AndGlobalModals({
 	let title = undefined;
 	switch (whichModal) {
 		case 'auth':
-			title = (
-				<FormattedMessage
-					id="andGlobalModals__heading"
-					defaultMessage="Member Only Feature"
-				/>
-			);
-			content = (
-				<>
-					<p>
-						<FormattedMessage
-							id="andGlobalModals__copy"
-							defaultMessage="You can bookmark audio to access later on your account. Other features include adding audio to your library."
+			title =
+				authState === 'register' ? (
+					<FormattedMessage
+						id="andGlobalModals__authRegisterTitle"
+						defaultMessage="Create Account"
+					/>
+				) : authState === 'login' ? (
+					<FormattedMessage
+						id="andGlobalModals__authLoginTitle"
+						defaultMessage="Login"
+					/>
+				) : (
+					<FormattedMessage
+						id="andGlobalModals__authTitle"
+						defaultMessage="Create a Free Account"
+					/>
+				);
+			content =
+				authState === 'register' ? (
+					<>
+						<SocialLogin
+							onSuccess={async () => {
+								handleClose();
+								onConfirm && onConfirm();
+								await refetchUserQueries(queryClient);
+							}}
 						/>
-					</p>
-					<div className={styles.buttonRow}>
-						<Button
-							type="super"
-							href={makeRegisterRoute(languageRoute)}
-							onClick={handleClose}
-							text={
-								<FormattedMessage
-									id="andGlobalModals__createAccount"
-									defaultMessage="Create Account"
-								/>
-							}
+						<RegisterForm
+							showLogin={() => setAuthState('login')}
+							onSuccess={async () => {
+								handleClose();
+								onConfirm && onConfirm();
+								await refetchUserQueries(queryClient);
+							}}
+							hideGuestButton
 						/>
-						<Button
-							type="primary"
-							href={makeLoginRoute(languageRoute)}
-							onClick={handleClose}
-							text={
-								<FormattedMessage
-									id="andGlobalModals__logIn"
-									defaultMessage="Log in"
-								/>
-							}
+					</>
+				) : authState === 'login' ? (
+					<>
+						<SocialLogin
+							onSuccess={async () => {
+								handleClose();
+								onConfirm && onConfirm();
+								await refetchUserQueries(queryClient);
+							}}
 						/>
-						<a
-							onClick={handleClose}
-							className={clsx(styles.continueAsGuest, 'decorated')}
-						>
+						<LoginForm
+							showRegister={() => setAuthState('register')}
+							onSuccess={() => {
+								handleClose();
+								onConfirm && onConfirm();
+							}}
+							hideGuestButton
+						/>
+					</>
+				) : (
+					<>
+						<p>
 							<FormattedMessage
-								id="andGlobalModals__continueAsGuest"
-								defaultMessage="Continue as guest"
+								id="andGlobalModals__authCopy"
+								defaultMessage="Create a free account or login to access the Library features, like saving, viewing your history, and syncing across devices."
 							/>
-						</a>
-					</div>
-				</>
-			);
+						</p>
+						<div className={styles.buttonRow}>
+							<Button
+								type="super"
+								onClick={() => setAuthState('register')}
+								text={
+									<FormattedMessage
+										id="andGlobalModals__createAccount"
+										defaultMessage="Create Account"
+									/>
+								}
+							/>
+							<Button
+								type="primary"
+								onClick={() => setAuthState('login')}
+								text={
+									<FormattedMessage
+										id="andGlobalModals__logIn"
+										defaultMessage="Log in"
+									/>
+								}
+							/>
+							<a
+								onClick={handleClose}
+								className={clsx(styles.continueAsGuest, 'decorated')}
+							>
+								<FormattedMessage
+									id="andGlobalModals__continueAsGuest"
+									defaultMessage="Continue as guest"
+								/>
+							</a>
+						</div>
+					</>
+				);
 			break;
 		case 'confirmRemoveFavorite':
 			title = (
