+++ src/containers/account/profile.tsx
@@ -8,7 +8,7 @@ import withAuthGuard from '@components/HOCs/withAuthGuard';
 import Button from '@components/molecules/button';
 import Input from '@components/molecules/form/input';
 import AccountNav from '@components/organisms/accountNav';
-import { invalidateAndResetUserQueries } from '@lib/api/login';
+import { refetchUserQueries } from '@lib/api/login';
 import {
 	useGetProfileDataQuery,
 	useUpdateProfileDataMutation,
@@ -26,7 +26,7 @@ function Profile(): JSX.Element {
 	const [surname, setSurname] = useState('');
 
 	const { mutate } = useUpdateProfileDataMutation({
-		onSuccess: () => invalidateAndResetUserQueries(queryClient),
+		onSuccess: () => refetchUserQueries(queryClient),
 	});
 	const intl = useIntl();
