+++ src/lib/api/useLogout.ts
@@ -2,11 +2,11 @@ import { useQueryClient } from 'react-query';
 
 import { clearSessionToken } from '@lib/cookies';
 
-import { invalidateAndResetUserQueries } from '.';
+import { resetUserQueries } from '.';
 
 export function useLogout(): Promise<void> {
 	const queryClient = useQueryClient();
 
 	clearSessionToken();
-	return invalidateAndResetUserQueries(queryClient);
+	return resetUserQueries(queryClient);
 }
